Отдельная таблица для Slave'ов. Исправил запись напряжения #2

+++ app/Device.php
@@ -64,7 +64,7 @@ class Device extends Model
      */
     protected $casts = [
         'address' => 'integer',
-        'voltage' => 'integer',
+        'voltage' => 'float',
         'alarm' => 'integer',
         'timeout' => 'integer',
         'sd_error' => 'datetime',
